refactor(components) : move mcq folder
