+++ src/components/quiz/mcq/reader/index.ts
@@ -1,3 +1,3 @@
-export { default as MCQView } from '../../MCQ_view';
+export { default as MCQView } from '../../../home/components/MCQ_view';
 export { default as SubmissionFeedback } from './parts/submission_feedback';
 export { default as useMCQReader } from './use_MCQ_reader';
