+++ src/components/quiz/mcq/reader/parts/hint_pane.tsx
