+++ src/components/quiz/mcq/shared/index.ts
